requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
billiard==4.2.0
bitarray==2.9.2
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
ckzg==2.0.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cytoolz==0.12.3
Django==5.1
django-celery-results==2.5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-jsonapi==7.0.2
drf-spectacular==0.27.2
eth-account==0.11.2
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==1.0.1
eth-typing==4.4.0
eth-utils==4.1.1
eth_abi==5.1.0
flower==2.0.1
frozenlist==1.4.1
hexbytes==0.3.1
hiredis==3.0.0
humanize==4.10.0
idna==3.7
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.0
lru-dict==1.2.0
multidict==6.0.5
parsimonious==0.10.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==5.27.3
psycopg2-binary==2.9.9
pycryptodome==3.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyunormalize==15.1.0
PyYAML==6.0.2
redis==5.0.8
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rlp==4.0.1
rpds-py==0.20.0
six==1.16.0
sqlparse==0.5.1
toolz==0.12.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
web3==6.20.2
websockets==13.0
yarl==1.9.4
